refactor(LottoPrize): LottoPrize를 data 패키지로 이동
LottoPrize를 의미에 맞게 data 패키지로 이동시켰습니다.
