+++ src/test/java/lotto/data/LottoResultTest.java
@@ -8,7 +8,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
 import lotto.domain.LottoResult;
-import lotto.message.LottoPrize;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
